Split Standard Integration into separate 'standard' folder.
