+++ demo-angular/app/demo/stripe.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from "@angular/core";
 import * as httpModule from "http";
-import { StripeAddress, StripeBackendAPI, StripeConfig, StripeCustomerSession, StripePaymentListener, StripePaymentSession, StripeShippingAddressField, StripeShippingMethod } from "nativescript-stripe";
+import { StripeAddress, StripeBackendAPI, StripeConfig, StripeCustomerSession, StripePaymentListener, StripePaymentSession, StripeShippingAddressField, StripeShippingMethod } from "nativescript-stripe/standard";
 import { Page } from "ui/page";
 
 // 1) To get started with this demo, first head to https://dashboard.stripe.com/account/apikeys
