+++ src/standard/index.ts
@@ -1,0 +1,1 @@
+export * from "./standard";
